requirements.txt
pyTelegramBotApi
pillow
opencv-python
pyaudio
psutil